requirements.txt
scapy
flask
flask-cors
requests
python-dotenv
